broke up preprocessing.py script in several submodules and refactored
