--- src/modules/data_processing/compose_transformations.py
@@ -1,8 +1,0 @@
-class Compose:
-    def __init__(self, transforms):
-        self.transforms = transforms
-
-    def __call__(self, sample):
-        for transform in self.transforms:
-            sample = transform(sample)
-        return sample
